2019-01-29/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# dev requirements
autopep8==1.6.0
pylint==2.12.2
nbconvert==6.4.4

# internal requirements
numpy==1.22.3
pandas==1.4.1
scikit-learn==1.0.2
matplotlib==3.5.1
